# JobFinder/src/main/java/com/ofdun/jobfinder/features/skill/api/dto/SkillDto.java
package com.ofdun.jobfinder.features.skill.api.dto;

import jakarta.validation.constraints.NotBlank;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class SkillDto {
    private Long id;

    @NotBlank(message = "Name is required")
    private String name;
}