# docs/swagger.yml
openapi: 3.0.3
info:
  title: JobFinder API
  description: API для платформы поиска работы с поддержкой векторного AI-мэтчинга резюме и вакансий.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Локальный сервер разработки
...
paths:
...
  /employers:
    post:
      tags: [Employers]
      summary: Регистрация нового работодателя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployerRequest'
      responses:
        '201':
          description: Работодатель создан
          content:
            application/json:
              schema:
                type: integer
  /employers/{id}:
    get:
      tags: [Employers]
      summary: Получить профиль работодателя
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Профиль работодателя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employer'
    put:
      tags: [Employers]
      summary: Обновить профиль работодателя
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployerRequest'
      responses:
        '200':
          description: Профиль обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employer'
    delete:
      tags: [Employers]
      summary: Удалить профиль работодателя
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Профиль успешно удален
...
  /resumes:
    post:
      tags: [Resumes]
      summary: Создать резюме
      description: Также асинхронно/синхронно генерирует векторные эмбеддинги (AI) и сохраняет в Qdrant.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResumeRequest'
      responses:
        '201':
          description: Резюме создано
          content:
            application/json:
              schema:
                type: integer
  /resumes/{id}:
    get:
      tags: [Resumes]
      summary: Получить резюме
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Данные резюме
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resume'
    put:
      tags: [Resumes]
      summary: Обновить резюме
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResumeUpdateRequest'
      responses:
        '200':
          description: Резюме обновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resume'
    delete:
      tags: [Resumes]
      summary: Удалить резюме
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Резюме удалено
...
  /resumes/search:
    get:
      tags: [Resumes]
      summary: Поиск резюме по фильтрам с пагинацией
      description: Параметры limit/offset/sortBy/sortDesc соответствуют модели OffsetPaginationParams.
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: Подстрока для поиска по описанию резюме
        - name: applicantId
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: categoryId
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: creationDateFrom
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: creationDateTo
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: skillIds
          in: query
          required: false
          schema:
            type: array
            items:
              type: integer
              format: int64
          style: form
          explode: true
          description: Фильтр по навыкам (AND). Резюме должно содержать ВСЕ переданные skillIds.
        - name: languageIds
          in: query
          required: false
          schema:
            type: array
            items:
              type: integer
              format: int64
          style: form
          explode: true
          description: Фильтр по языкам (AND). Резюме должно содержать ВСЕ переданные languageIds.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 20
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: sortBy
          in: query
          required: false
          schema:
            type: string
        - name: sortDesc
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Страница резюме
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumePageResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Location:
      type: object
      properties:
        id:
          type: integer
        city:
          type: string
        country:
          type: string
    Skill:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
    Language:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        proficiencyLevel:
          type: string
          enum: [A1, A2, B1, B2, C1, C2]
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
    Education:
      type: object
      properties:
        id:
          type: integer
          description: ID записи об образовании (не передаётся при создании внутри POST /resumes)
        resumeId:
          type: integer
          description: ID резюме (не передаётся при создании внутри POST /resumes, проставляется сервером)
        educationDegree:
          type: string
          enum: [BACHELOR, MASTER, SPECIALIST, PHD]
        institutionName:
          type: string
        faculty:
          type: string
        department:
          type: string
        yearOfGraduation:
          type: string
          format: date
    EducationCreate:
      type: object
      description: Образование для создания вместе с резюме (в POST /resumes поля id и resumeId не передаются)
      properties:
        educationDegree:
          type: string
          enum: [BACHELOR, MASTER, SPECIALIST, PHD]
        institutionName:
          type: string
        faculty:
          type: string
        department:
          type: string
        yearOfGraduation:
          type: string
          format: date
    JobExperience:
      type: object
      properties:
        id:
          type: integer
          description: ID записи опыта работы (не передаётся при создании внутри POST /resumes)
        resumeId:
          type: integer
          description: ID резюме (не передаётся при создании внутри POST /resumes, проставляется сервером)
        position:
          type: string
        companyName:
          type: string
        description:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
    JobExperienceCreate:
      type: object
      description: Опыт работы для создания вместе с резюме (в POST /resumes поля id и resumeId не передаются)
      properties:
        position:
          type: string
        companyName:
          type: string
        description:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
...
    Employer:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          writeOnly: true
        description:
          type: string
        address:
          type: string
        websiteLink:
          type: string
          format: uri
        location:
          $ref: '#/components/schemas/Location'
...
    Resume:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        applicantId:
          type: integer
        category:
          $ref: '#/components/schemas/Category'
        description:
          type: string
        skills:
          type: array
          items:
            $ref: '#/components/schemas/Skill'
        educations:
          type: array
          items:
            $ref: '#/components/schemas/Education'
        jobExperiences:
          type: array
          items:
            $ref: '#/components/schemas/JobExperience'
        languages:
          type: array
          items:
            $ref: '#/components/schemas/Language'
        date:
          type: string
          format: date-time
          readOnly: true
...
    EmployerRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          writeOnly: true
        description:
          type: string
        address:
          type: string
        websiteLink:
          type: string
          format: uri
        locationId:
          type: integer
...
    ResumeRequest:
      type: object
      properties:
        applicantId:
          type: integer
        categoryId:
          type: integer
        description:
          type: string
        skillIds:
          type: array
          items:
            type: integer
        educations:
          type: array
          items:
            $ref: '#/components/schemas/EducationCreate'
        jobExperiences:
          type: array
          items:
            $ref: '#/components/schemas/JobExperienceCreate'
        languageIds:
          type: array
          items:
            type: integer
    ResumeUpdateRequest:
      type: object
      description: Запрос на обновление резюме. Поля опциональны.
      properties:
        applicantId:
          type: integer
        categoryId:
          type: integer
        description:
          type: string
        skillIds:
          type: array
          items:
            type: integer
        educations:
          type: array
          items:
            $ref: '#/components/schemas/EducationCreate'
        jobExperiences:
          type: array
          items:
            $ref: '#/components/schemas/JobExperienceCreate'
        languageIds:
          type: array
          items:
            type: integer
...
    ResumePageResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Resume'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
...